Version 0.10 "SuperConsoleManager"
Added:
1. Added Main.java;
2. Added Console managers;
3. Added Controllers to all database repositories;
4. Other little improvements..

+++ src/main/java/controllers/AdminController.java
@@ -26,37 +26,52 @@ public class AdminController {
     }
 
     public String updateAdmin(int id, String columnName, Object value) {
-        if (adminRepository.getById(id) != null) {
+        if (checkAdminExistence(id)) {
             boolean updated = adminRepository.updateRecord(id, columnName, value);
 
             if (updated)
-                return "\nAdmin's information with id " + id + " in " + columnName + " updated successfully.";
+                return "\nAdmin's information with ID " + id + " in " + columnName + " updated successfully.";
             else return "\nUpdate was failed!";
 
         } else return "\nAdmin with ID " + id + " not found.";
     }
     
     public void deleteAdmin(String input) {
-        String[] adminIdsString = input.split(",");
+        String regex1 = "^\\d+(, \\d+)*$"; // For case if user enters something like "27, 28". Separated with comma and space
+        String regex2 = "^\\d+(,\\d+)*$"; // For case if user enters something like "27,28". Separated only comma.
+         /*
+         I searched this in StackOverFlow. Please forgive me. (Zhasulan)
+                ^ - start of the string
+                \\d+ - one or more digits
+                (,\\d+)* - a comma followed by one or more digits, repeated zero or more times
+                $ - end of the string
+         */
 
-        // I searched converting String to "int..." type in StackOverFlow. Please, do not think that this is copy-paste.
-        // I can easily explain how this code works.
-        // (Zhasulan)
-
-        int[] adminIds = new int[adminIdsString.length];
-        for (int i = 0; i < adminIdsString.length; i++) {
-            adminIds[i] = Integer.parseInt(adminIdsString[i].trim());
+        if (!input.matches(regex1) & !input.matches(regex2)) {
+            System.err.println("\nIncorrect input. Enter IDs (or one ID) separated by comma.");
         }
+        else {
+            String[] adminIdsString = input.split(",");
 
-        for (int adminId : adminIds) {
-            if(checkAdminExistence(adminId)) {
+            // I searched converting String to "int..." type in StackOverFlow. Please, do not think that this is copy-paste.
+            // I can easily explain how this code works.
+            // (Zhasulan)
 
-                boolean deleted = adminRepository.deleteRecord(adminId);
-                if (deleted)
-                    System.out.println("\nAdmin with ID " + adminId + " deleted successfully.");
-                else System.out.println("\nDelete of admin with ID " + adminId + " was failed!");
+            int[] adminIds = new int[adminIdsString.length];
+            for (int i = 0; i < adminIdsString.length; i++) {
+                adminIds[i] = Integer.parseInt(adminIdsString[i].trim());
+            }
 
-            } else System.out.println("\nAdmin with ID " + adminId + " not found.");
+            for (int adminId : adminIds) {
+                if (checkAdminExistence(adminId)) {
+
+                    boolean deleted = adminRepository.deleteRecord(adminId);
+                    if (deleted)
+                        System.out.println("\nAdmin with ID " + adminId + " deleted successfully.");
+                    else System.out.println("\nDelete of admin with ID " + adminId + " was failed!");
+
+                } else System.out.println("\nAdmin with ID " + adminId + " not found.");
+            }
         }
     }
